분석(20260727_ble-protocol-restructure)과 8비트 전환
(20260727_ble-8bit-packet-redesign)을 함께 병합한다.

분석
- 복잡도 근인 3건 규명(함수 분해 부재·라이브 상태 4층 분산·파서가
  자극 FSM 구동), 라이브모드 상태머신 사양화, 재구조화 6단계 로드맵
- 참고 문서 13종 신설 (docs/참고/ble/)

8비트 전환 (B1~B3c)
- SPI1 은 이미 8비트 워드인데 DMA 만 32비트라 패킷 버퍼가 int[] 였다.
  WORD_SIZE_8BITS_TO_8BITS 로 전환해 uint8_t[] 로 1:1 대응시켰다.
- 정적 메모리 186B, 코드 1,448B 절감. 잔여 int 조립 버퍼 26 -> 0
- 전 단계 빌드 무결, 실기 5단계 정상 확인

전환 중 발견한 기존 결함 2건 수정
- 바이폴라 기준전극 electrodeMap 범위 밖 읽기
- EEPROM 슬롯 0 허용으로 인한 map[-1] 접근 (심층 방어)

+++ src/2__cm3/source/ble/tdc_ble_gain_control.c
@@ -31,7 +31,7 @@ void tdc_ble_gain_control_init(void)
     cfx_cm3_sharedMemoryAll.is_i2s_source_cradle = 0;
 }
 
-int tdc_ble_gain_control_handle(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index)
+int tdc_ble_gain_control_handle(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index)
 {
     int control_type = packet->data[0];  // 0 = Read, 1 = Write
     int gain_type    = packet->data[1];  // 1 = Gain_A, 2 = Gain_B
