+++ src/2__cm3/source/ble/tdc_ble_gain_control.h
@@ -49,6 +49,6 @@ void tdc_ble_gain_control_init(void);
  *  - tx_index : 현재 쓰기 오프셋(진입 시점)
  *  - 반환값   : 응답을 채운 뒤의 tx_index
  */
-int tdc_ble_gain_control_handle(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index);
+int tdc_ble_gain_control_handle(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index);
 
 #endif /* __tdc_ble_gain_control_h__ */
